Move tags commands to their own file. Completely removed functions.py. Made some class methods static methods as there was no reason for them to be class methods.
