+++ modules/commands.py
@@ -1,65 +1,16 @@
-import modules.functions as func
 import botoptions
 import asyncio
 import discord
 import os
 import random
+import config
+import aiohttp
 from discord.ext import commands
 
 
 class Commands:
     def __init__(self, bot):
         self.bot = bot
-        self.bot.database.create_table('''CREATE TABLE IF NOT EXISTS storage(label TEXT PRIMARY KEY unique, author TEXT, msg TEXT)''')
-
-    # TODO: Move the Tags commands into a separate file
-    @commands.group(aliases=['tags'])
-    async def Tags(self, ctx):
-        """-Command group for all storage commands"""
-        if ctx.invoked_subcommand is None:
-            await ctx.channel.send('You need to pass a subcommand. Type $eli help storage for more info.')
-
-    @Tags.command(aliases=['tag', 'set'])
-    async def Tag(self, ctx, label: str, *, msg: str):
-        """-Stores a message
-        $eli tags tag label msg
-        label: This must not contain spaces, use _ to represent spaces This_Is_An_Example
-        msg: Can be as long as or how ever many lines you want"""
-        msg = await self.bot.database.insert_tag_data(label, str(ctx.author), msg)
-        await ctx.channel.send(msg)
-
-    @Tags.command(aliases=['update'])
-    async def UpdateMessage(self, ctx, label: str, *, msg: str):
-        """-Updates the message contained in a specific label"""
-        await self.bot.database.update_data(label, msg)
-        await ctx.channel.send(label+' updated!')
-
-    @Tags.command(aliases=['retrieve', 'get'])
-    async def RetrieveMessage(self, ctx, label):
-        """-Retrieves a message
-        label: The name of the message to retrieve"""
-        msg = await self.bot.database.retrieve_data(label)
-        await ctx.channel.send('```'
-                               f'Author: {msg[1]}\n'
-                               f'{msg[0]}'
-                               '```')
-
-    @Tags.command(aliases=['remove', 'delete'])
-    async def RemoveMessage(self, ctx, label):
-        """-Removes a message
-        label: The name of the message to delete"""
-        await self.bot.database.delete_data(label)
-        await ctx.channel.send('Removed stored message with the label ' + label)
-
-    @Tags.command(aliases=['listall', 'listmessages'])
-    async def ListMessages(self, ctx):
-        """-Lists all saved messages"""
-        msg = await self.bot.database.retrieve_all_labels()
-        if msg[1] is True:
-            message = "\n".join(str(i) for i in msg[0])
-            await ctx.channel.send(message)
-        else:
-            await ctx.channel.send(msg[0])
 
     @commands.command(aliases=['raidtime'])
     async def RaidTime(self, ctx):
@@ -84,7 +35,7 @@ class Commands:
     async def TopClip(self, ctx):
         """-It's the top clip from my channel!"""
         await ctx.channel.send("Check out this amazing clip from my channel!")
-        await ctx.channel.send(await func.RetrieveTwitchClip('elimere'))
+        await ctx.channel.send(await self.retrieve_twitch_clip('elimere'))
 
     @commands.command(aliases=['raidmods'])
     async def RaidMods(self, ctx):
@@ -172,6 +123,27 @@ class Commands:
                     await ctx.channel.send(response.content)  # Else just send what the bot response was
         except asyncio.TimeoutError:  # This fires if the user doesn't respond in 20 seconds
             await ctx.channel.send("I guess you didn't have anything to say anyways....")
+
+    @staticmethod
+    async def retrieve_twitch_clip(channel):
+        """Retrieves the newest twitch clip from the channel"""
+        # Docs located here https://dev.twitch.tv/docs/v5/reference/clips
+        twitchURL = 'https://api.twitch.tv/kraken/clips/top'
+        headers = {
+            'Client-ID': config.twitchBotId,
+            'Accept': 'application/vnd.twitchtv.v5+json'
+        }
+        params = {
+            'channel': channel,
+            'period': 'all',
+            'limit': '1'
+        }
+        async with aiohttp.ClientSession() as session:
+            async with session.get(twitchURL, headers=headers, params=params) as resp:
+                json_info = await resp.json()
+                await asyncio.sleep(0.250)
+                session.close()
+        return json_info['clips'][0]['url']
 
 
 def setup(bot):
